added validation and ui changes in add task ui
